Removed int64 restriction on torrent.File.Length

+++ torrent/file.go
@@ -6,7 +6,7 @@ import (
 
 type File struct {
 	PathComponents []string `bencode:"path"`
-	Length         int64    `bencode:"length"`
+	Length         int      `bencode:"length"`
 	MD5sum         string   `bencode:"md5sum"`
 }
